Refactor classes and packages
