+++ src/main/java/net/vexelon/currencybg/app/ui/components/WalletListAdapter.java
@@ -34,7 +34,7 @@ import net.vexelon.currencybg.app.R;
 import net.vexelon.currencybg.app.common.Sources;
 import net.vexelon.currencybg.app.db.models.CurrencyData;
 import net.vexelon.currencybg.app.db.models.WalletEntry;
-import net.vexelon.currencybg.app.ui.UIUtils;
+import net.vexelon.currencybg.app.ui.utils.UIUtils;
 import net.vexelon.currencybg.app.utils.NumberUtils;
 
 import org.joda.time.LocalDateTime;
